# Lazy-load pandas