@bennybp 2024-11-12
Sounds good to me.

I think the CI errors are from conda resolving some older pytest version with python 3.8. Since python 3.8 is now EOL, I can probably take that out of the CI. Would you agree?