@mattwthompson 2024-11-12
Sure, I've made an attempt in https://github.com/MolSSI/QCFractal/pull/856