@bennybp 2024-11-12
Thanks! This is great, I'm always looking for performance increases